+++ concon/cli.py
@@ -17,9 +17,9 @@ from .core import ConConError
 DEFAULT_CONFIG = 'config/config.yml'
 DEFAULT_LOG_CONFIG = 'config/logging_global.cfg'
 
-
 # logging.basicConfig(level=logging.DEBUG)
-logger = logging.getLogger('UCA console')
+logger = logging.getLogger('ConCon')
+
 
 
 def report_errors(err):
@@ -34,13 +34,25 @@ def report_errors(err):
               help="Application specific configuration file.")
 @click.option('-d', '--device', default=None,
               help="Selected device to configure.")
+@click.option('--verbose/--quiet', default=False,
+              help="Enable / disable logging to Log.txt file")
 @click.pass_context
-def main(ctx, config, device, args=None):
+def main(ctx, config, device, verbose, args=None):
     """Tool for configuration of devices implementing "Uniprot" communication
     layer over USB (HID profile).
     """
     ctx.obj = {}
     logging.basicConfig(level=logging.ERROR)
+    
+    # Check if verbose mode is on
+    if(verbose):
+        logging.config.fileConfig(
+        pkg_resources.resource_filename('concon',
+                                        DEFAULT_LOG_CONFIG),
+        None,
+        False)
+    # /if verbose
+    
     # TODO: Config doesn't work...
     # logging.config.fileConfig(
     #     pkg_resources.resource_filename('concon', DEFAULT_LOG_CONFIG)
@@ -89,7 +101,7 @@ def main(ctx, config, device, args=None):
             # Show device selection prompt
             click.secho("More than one configurable devices detected.\n"
                         "Choose one from below and run concon with"
-                        " -d [0 ~ {0}] option.".format(len(found_devices)-1),
+                        " -d [0 ~ {0}] option.".format(len(found_devices) - 1),
                         fg='yellow')
             ctx.invoke(device_list)
             raise click.Abort()
@@ -99,10 +111,9 @@ def main(ctx, config, device, args=None):
         except (IndexError, ValueError):
             report_errors('Invalid device number "{0}". '
                           'Choose between 0~{1}'.format(device,
-                                                        len(found_devices)-1))
+                                                        len(found_devices) - 1))
             ctx.invoke(device_list)
             raise click.Abort()
-
 
 
 @main.command()
@@ -124,7 +135,7 @@ def write(ctx, file_name):
         click.echo("Reading configuration from: {0}".format(file_name))
         cfg_pars.read_setting_from_file(file_name,
                                         ignore_errors=False,
-                                        try_fix_errors=True)
+                                        try_fix_errors=False)
 
         # And send changes to device
         label = "Writing configuration to {0}".format(device.name)
@@ -176,6 +187,7 @@ def device_list(ctx):
     for i in range(len(found_devices)):
         click.echo(" Dev# {0} <{1}>".format(
             i, found_devices[i].name))
+
 
 if __name__ == "__main__":
     main()