+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.9.8 (2018-05-04)
+------------------
+* Added information how to access device s regular user (sudo not needed)
+* Added verbose option
+* When communication with device will fail, more friendly message is shown
+* Updated links in documentation
+
 0.9.0 (2017-06-13)
 ------------------